meta-facebook: minerva: Modify the fan status monitor
The minerva-fan-status-monitor monitors all fan statuses on the system and sends a log with the CALLOUT INVENTORY PATH to instruct the phosphor-led-manager to light the FAN STATUS LED on the front of the CMM.

We have found that on some systems, logs are recorded but the status is not retained, which makes debugging difficult. Therefore, we have changed the log message to include the "reason" to help users identify the problem more easily.

Change-Id: Ibed1e04b773e91a6aa70cb4e184d6a8550570a56
Signed-off-by: Yang Chen <yang.chen@quantatw.com>